+++ backend/src/main/java/dsi/ruet/backend/dto/mealmanager/DailyTopUpSummary.java
@@ -4,13 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class DailyTopUpSummary {
     private String date;
-    private BigDecimal totalAmount;
+    private Long totalAmount;
     private long transactionCount;
 }
