+++ backend/src/main/java/dsi/ruet/backend/dto/mealmanager/TopUpRequest.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 /**
  * Request body for wallet top-up.
  * studentId is the student's roll number (String), e.g. "S2021001".
@@ -15,5 +13,5 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 public class TopUpRequest {
     private String studentId;     // student roll number (matches frontend)
-    private BigDecimal amount;    // amount to add (1 credit = 1 BDT)
+    private Long amount;          // amount to add (1 credit = 1 BDT)
 }
